usb: tcpm: fix toggling in host (SRC) mode
PU\PD resistors on CC lines must be configured before running the TCPM
state machine.

Also, when the Type-C port acts as a host (SRC), the VBUS sould be enabled
only after the toggling has been completed. And we have to wait for
the corresponding IRQ to finish the toggling process. But this doesn't
happen because we exit from tcpm_poll_event() if VBUS != OK.

So, Let's check for VBUS state only when the port acts as a host (SRC)
to solve this problem.

Signed-off-by: Maksim Kiselev <bigunclemax@gmail.com>
Link: https://lore.kernel.org/r/20241101103501.21432-4-bigunclemax@gmail.com
Signed-off-by: Sebastian Reichel <sre@ring0.de>